WeepWand adjustments

+++ Content/Projectiles/Friendly/Mage/WeepWandWisp.cs
@@ -17,6 +17,7 @@ namespace ITD.Content.Projectiles.Friendly.Mage
         {
             Main.projFrames[Projectile.type] = 8;
         }
+
         public override void SetDefaults()
         {
             Projectile.DamageType = DamageClass.Magic;
@@ -30,11 +31,12 @@ namespace ITD.Content.Projectiles.Friendly.Mage
             Projectile.tileCollide = false;
 
             Projectile.usesLocalNPCImmunity = true;
-            Projectile.localNPCHitCooldown = (15);
+            Projectile.localNPCHitCooldown = 15;
 
             Projectile.frame = 0;
             Projectile.frameCounter = 0;
         }
+
         private NPC HomingTarget
         {
             get => Projectile.ai[0] == 0 ? null : Main.npc[(int)Projectile.ai[0] - 1];
@@ -43,6 +45,7 @@ namespace ITD.Content.Projectiles.Friendly.Mage
                 Projectile.ai[0] = value == null ? 0 : value.whoAmI + 1;
             }
         }
+
         public override void AI()
         {
             float maxDetectRadius = 800f;
@@ -62,52 +65,61 @@ namespace ITD.Content.Projectiles.Friendly.Mage
                 return;
             }
 
-            Projectile.friendly = true;
-            Projectile.velocity *= 1.05f;
-
             if (Projectile.penetrate > 0)
             {
                 HomingTarget ??= Projectile.FindClosestNPC(maxDetectRadius);
 
                 if (HomingTarget == null)
                 {
-                    Projectile.velocity *= 0.90f;
+                    Projectile.velocity *= 0.85f;
                     deathTimer++;
                     if (deathTimer > 45)
                     {
                         for (int i = 0; i < 6; i++)
                         {
-                            Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, ModContent.DustType<WispDust>(), 0f, 0f, 150, default(Color), 1f);
+                            Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustID.UltraBrightTorch, 0f, 0f, 150, default(Color), 0.25f);
                         }
                         Projectile.Kill();
                     }
                     return;
                 }
-
-                Vector2 directionToTarget = HomingTarget.Center - Projectile.Center;
-
-                float targetAngle = directionToTarget.ToRotation();
-
-                float currentAngle = Projectile.velocity.ToRotation();
-
-                float angleStep = MathHelper.ToRadians(2);
-                float newAngle = currentAngle.AngleTowards(targetAngle, angleStep);
-
-                float length = Projectile.velocity.Length();
-                Projectile.velocity = newAngle.ToRotationVector2() * length;
-
-                Projectile.Center += Main.rand.NextVector2Circular(0.5f, 0.5f); 
-
-                Projectile.rotation = newAngle + MathHelper.PiOver2;
-            }
-            else
-            {
-                for (int i = 0; i < 6; i++)
+                
+                if (IsValidTarget(HomingTarget))
                 {
-                    Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, ModContent.DustType<WispDust>(), 0f, 0f, 150, default(Color), 1f);
+
+                    Vector2 directionToTarget = HomingTarget.Center - Projectile.Center;
+                    directionToTarget.Normalize();
+                    float speed = 10f;
+                    Projectile.velocity = Vector2.Lerp(Projectile.velocity, directionToTarget * speed, 0.1f);
+
+                    float targetAngle = directionToTarget.ToRotation();
+                    float currentAngle = Projectile.velocity.ToRotation();
+                    float angleStep = MathHelper.ToRadians(4);
+                    float newAngle = currentAngle.AngleTowards(targetAngle, angleStep);
+
+                    float length = Projectile.velocity.Length();
+                    Projectile.velocity = newAngle.ToRotationVector2() * length;
+
+                    Projectile.Center += Main.rand.NextVector2Circular(0.5f, 0.5f);
+
+                    Projectile.rotation = newAngle + MathHelper.PiOver2;
+
+                    Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustID.UltraBrightTorch, 0f, 0f, 150, default(Color), 0.25f);
                 }
-                Projectile.Kill();
+                else
+                {
+                    for (int i = 0; i < 6; i++)
+                    {
+                        Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustID.UltraBrightTorch, 0f, 0f, 150, default(Color), 0.25f);
+                    }
+                    Projectile.Kill();
+                }
             }
+        }
+
+        public bool IsValidTarget(NPC target)
+        {
+            return target.CanBeChasedBy() && Collision.CanHit(Projectile.Center, 1, 1, target.position, target.width, target.height);
         }
     }
 }
